dom/tests/mochitest/mouselock/test_TargetOutOfFocus.html

<!DOCTYPE HTML>
<html>
<!--
Testing to ensure that mouse-lock still applies when target moves out of focus.
-->
<head>
  <title>Test for Infocus</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>

<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=633602">Mozilla Bug 633602</a>
<div id="message">Setting Mouse-Lock while document is not in focus.</div><br />
<canvas id="canvas" width="150" height="150"></canvas>

<script type="text/javascript">
SimpleTest.waitForExplicitFinish();
var windowObjectReference = null; // global variable
var pointer = navigator.pointer;
var canvas = document.getElementById("canvas");

canvas.mozRequestFullScreen();

  document.addEventListener("mozfullscreenchange", function (e) {
        if (document.mozFullScreen) {
          if (document.mozFullScreenElement === canvas) {
              // Mouse should not be lock because the window is not in focus
            if(windowObjectReference == null || windowObjectReference.closed || )
               /* if the pointer to the window object in memory does not exist
                  or if such pointer exists but the window was closed */
            {
               pointer.lock(canvas);

               windowObjectReference = window.open("http://www.spreadfirefox.com/",
              "PromoteFirefoxWindowName", "resizable=yes,scrollbars=yes,status=yes");
                 /* then create it. The new window will be created and
                    will be brought on top of any other window. */
            }
            window.blur();//Sets the main window out of focus.
            is(pointer.islocked(), true, "Mouse pointer should be locked");


            window.focus();
            is(pointer.islocked(), true, "Mouse should be locked after regaining focus");
          }

          SimpleTest.finish();
        }
        else {
            canvas.mozRequestFullScreen();
        }
  }, false);

</script>
</body>
</html>
